refactor: export file converted into ts
